community/2025: move plotting functions from process to altplot
